+++ packages/picasso-codemod/src/v5.0.0/label-tag/__testfixtures__/import-as.input.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import { Label as Something } from '@toptal/picasso'
+
+const DefaultExample = () => <Something>Content</Something>
+
+export default DefaultExample