+++ packages/picasso-codemod/src/v5.0.0/label-tag/__tests__/test.ts
@@ -1,6 +1,6 @@
 import { defineTest } from 'jscodeshift/src/testUtils'
 
-const testTypes = ['basic', 'group']
+const testTypes = ['basic', 'group', 'import-as']
 
 testTypes.forEach(testType => {
   defineTest(__dirname, 'label-tag', {}, testType, { parser: 'tsx' })